move white into community themes
